requirements.txt
opencv-python
numpy
tensorflow
keras
scikit-learn
pyttsx3         # (Optional, for text-to-speech)
Pillow
streamlit       # (Optional, for GUI)
